physics-experiment-helper/pyproject.toml
[project]
name = "physics-experiment-helper"
version = "1.0.0"
description = "AI-powered physics experiment guide generator for Grade 9 students"
requires-python = ">=3.10"
dependencies = [
    "deepagents>=0.1.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.20",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.3.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.8.0",
    "markdown",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

[tool.setuptools]
packages = ["agent", "api"]
package-dir = {"" = "src"}
